renamed 'rng command' to 'choose command' and implemented 'helper' class for general purpose functions
